Add ability to sort list by phone number
The address book is currently unable to sort list by phone number.

The user may find it more convenient to work with a list in which the
phone numbers are sorted.

Enhance sort method to sort the list by phone number.

Users can thus easily find contacts with a spefic phone number.